Fix integration test depending on order of results
Move assert for vector comparison into separate module. Add assert for
MeTTa runner results comparison.
