+++ lib/src/common/collections.rs
@@ -135,26 +135,3 @@ impl Display for ImmutableString {
     }
 }
 
-pub fn vec_eq_some_order<T: PartialEq + std::fmt::Debug>(left: &Vec<T>, right: &Vec<T>) -> bool {
-    let mut amap: ListMap<&T, usize> = ListMap::new();
-    let mut bmap: ListMap<&T, usize> = ListMap::new();
-    for i in left.iter() {
-        *amap.entry(&i).or_insert(0) += 1;
-    }
-    for i in right.iter() {
-        *bmap.entry(&i).or_insert(0) += 1;
-    }
-    amap == bmap
-}
-
-#[macro_export]
-macro_rules! assert_eq_no_order {
-    ($left:expr, $right:expr) => {
-        {
-            let left = &$left;
-            let right = &$right;
-            assert!($crate::common::collections::vec_eq_some_order(left, right),
-                "(left == right some order)\n  left: {:?}\n right: {:?}", left, right);
-        }
-    }
-}
